pyproject.toml
[project]
name = "c2md"
version = "0.1.0"
description = "Content to Markdown - convert URLs and files to clean markdown"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.49",
    "markitdown>=0.1",
    "markdownify>=0.14",
    "httpx>=0.24",
    "readability-lxml>=0.8",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "click>=8.0",
    "rich>=13.0",
    "pillow>=10.0",
    "xxhash>=3.0",
    "orjson>=3.9",
]

[project.scripts]
c2md = "c2md.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0",
]
